More Audio Code
Far more audio code as well as some fixed errors (which is always good).

Don't you hate when you coded so long that you keep typing sentences ending with semicolins?

+++ Leviathan/Input.h
@@ -9,7 +9,7 @@ class Input {
 
 public:
 
-	int delta; //Time ratio (delta)
+	double delta; //Time ratio (delta)
 
 	Input();
 	~Input();
